+++ src/main/java/com/powsybl/openloadflow/OpenLoadFlowProvider.java
@@ -27,10 +27,7 @@ import com.powsybl.openloadflow.dc.DcLoadFlowEngine;
 import com.powsybl.openloadflow.dc.DcLoadFlowResult;
 import com.powsybl.openloadflow.graph.EvenShiloachGraphDecrementalConnectivityFactory;
 import com.powsybl.openloadflow.graph.GraphConnectivityFactory;
-import com.powsybl.openloadflow.network.LfBranch;
-import com.powsybl.openloadflow.network.LfBus;
-import com.powsybl.openloadflow.network.LfNetwork;
-import com.powsybl.openloadflow.network.NominalVoltageMapping;
+import com.powsybl.openloadflow.network.*;
 import com.powsybl.openloadflow.network.impl.LfNetworkLoaderImpl;
 import com.powsybl.openloadflow.network.impl.Networks;
 import com.powsybl.openloadflow.network.util.ZeroImpedanceFlows;
@@ -98,7 +95,7 @@ public class OpenLoadFlowProvider implements LoadFlowProvider {
             LOGGER.info("Outer loops: {}", acParameters.getOuterLoops().stream().map(OuterLoop::getType).collect(Collectors.toList()));
         }
 
-        NominalVoltageMapping nominalVoltageMapping = NominalVoltageMapping.create(network, parametersExt.getNominalVoltagePerUnitResolution());
+        NominalVoltageMapping nominalVoltageMapping = SimpleNominalVoltageMapping.create(network, parametersExt.getNominalVoltagePerUnitResolution());
 
         List<AcLoadFlowResult> results;
         if (parametersExt.isNetworkCacheEnabled()) {
@@ -174,7 +171,7 @@ public class OpenLoadFlowProvider implements LoadFlowProvider {
 
         var dcParameters = OpenLoadFlowParameters.createDcParameters(network, parameters, parametersExt, matrixFactory, connectivityFactory, forcePhaseControlOffAndAddAngle1Var);
 
-        NominalVoltageMapping nominalVoltageMapping = NominalVoltageMapping.create(network, parametersExt.getNominalVoltagePerUnitResolution());
+        NominalVoltageMapping nominalVoltageMapping = SimpleNominalVoltageMapping.create(network, parametersExt.getNominalVoltagePerUnitResolution());
 
         List<DcLoadFlowResult> results = DcLoadFlowEngine.run(network, nominalVoltageMapping, new LfNetworkLoaderImpl(), dcParameters, reporter);
