+++ src/main/java/com/powsybl/openloadflow/network/LfNetwork.java
@@ -480,15 +480,15 @@ public class LfNetwork extends AbstractPropertyBag implements PropertyBag {
     }
 
     public static <T> List<LfNetwork> load(T network, LfNetworkLoader<T> networkLoader, SlackBusSelector slackBusSelector) {
-        return load(network, NominalVoltageMapping.NONE, networkLoader, new LfNetworkParameters().setSlackBusSelector(slackBusSelector), Reporter.NO_OP);
+        return load(network, SimpleNominalVoltageMapping.NONE, networkLoader, new LfNetworkParameters().setSlackBusSelector(slackBusSelector), Reporter.NO_OP);
     }
 
     public static <T> List<LfNetwork> load(T network, LfNetworkLoader<T> networkLoader, LfNetworkParameters parameters) {
-        return load(network, NominalVoltageMapping.NONE, networkLoader, parameters, Reporter.NO_OP);
+        return load(network, SimpleNominalVoltageMapping.NONE, networkLoader, parameters, Reporter.NO_OP);
     }
 
     public static <T> List<LfNetwork> load(T network, LfNetworkLoader<T> networkLoader, SlackBusSelector slackBusSelector, Reporter reporter) {
-        return load(network, NominalVoltageMapping.NONE, networkLoader, new LfNetworkParameters().setSlackBusSelector(slackBusSelector), reporter);
+        return load(network, SimpleNominalVoltageMapping.NONE, networkLoader, new LfNetworkParameters().setSlackBusSelector(slackBusSelector), reporter);
     }
 
     public static <T> List<LfNetwork> load(T network, NominalVoltageMapping nominalVoltageMapping, LfNetworkLoader<T> networkLoader, LfNetworkParameters parameters, Reporter reporter) {
