+++ src/test/java/com/powsybl/openloadflow/sa/LfContingencyTest.java
@@ -74,7 +74,7 @@ class LfContingencyTest extends AbstractConverterTest {
         String branchId = "LINE_S3S4";
         Contingency contingency = new Contingency(branchId, new BranchContingency(branchId));
         List<PropagatedContingency> propagatedContingencies =
-            PropagatedContingency.createList(network, Collections.singletonList(contingency), new HashSet<>(), false, false, false, true, NominalVoltageMapping.NONE);
+            PropagatedContingency.createList(network, Collections.singletonList(contingency), new HashSet<>(), false, false, false, true, SimpleNominalVoltageMapping.NONE);
 
         List<LfContingency> lfContingencies = propagatedContingencies.stream()
                 .flatMap(propagatedContingency -> propagatedContingency.toLfContingency(mainNetwork).stream())
@@ -103,7 +103,7 @@ class LfContingencyTest extends AbstractConverterTest {
         String generatorId = "GEN";
         Contingency contingency = new Contingency(generatorId, new GeneratorContingency(generatorId));
         assertThrows(PowsyblException.class, () ->
-                        PropagatedContingency.createList(network, Collections.singletonList(contingency), new HashSet<>(), false, false, false, true, NominalVoltageMapping.NONE),
+                        PropagatedContingency.createList(network, Collections.singletonList(contingency), new HashSet<>(), false, false, false, true, SimpleNominalVoltageMapping.NONE),
                 "Generator 'GEN' not found in the network");
     }
 
@@ -119,7 +119,7 @@ class LfContingencyTest extends AbstractConverterTest {
         String loadId = "LOAD";
         Contingency contingency = new Contingency(loadId, new LoadContingency(loadId));
         assertThrows(PowsyblException.class, () ->
-                        PropagatedContingency.createList(network, Collections.singletonList(contingency), new HashSet<>(), false, false, false, true, NominalVoltageMapping.NONE),
+                        PropagatedContingency.createList(network, Collections.singletonList(contingency), new HashSet<>(), false, false, false, true, SimpleNominalVoltageMapping.NONE),
                 "Load 'LOAD' not found in the network");
     }
 }
