+++ src/test/java/com/powsybl/openloadflow/sensi/dc/DcSensitivityAnalysisTest.java
@@ -829,7 +829,7 @@ class DcSensitivityAnalysisTest extends AbstractSensitivityAnalysisTest {
         List<Contingency> contingencies = List.of(new Contingency("c1", new BranchContingency("L1")));
 
         List<PropagatedContingency> propagatedContingencies = PropagatedContingency.createList(network, contingencies,
-                Collections.emptySet(), false, false, false, false, NominalVoltageMapping.NONE);
+                Collections.emptySet(), false, false, false, false, SimpleNominalVoltageMapping.NONE);
         assertEquals(1, propagatedContingencies.size());
     }
